+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add Admin note field to the Instance field mapping profile: Create/Edit (UIDATIMP-1118)
 * Add Admin note field to the Holdings field mapping profile: Create/Edit (UIDATIMP-1119)
 * Add Admin note field to the Instance field mapping profile: View (UIDATIMP-1121)
+* Add Admin note field to the Item field mapping profile: Create/Edit (UIDATIMP-1120)
 
 ## [5.1.2](https://github.com/folio-org/ui-data-import/tree/v5.1.2) (2022-03-24)